requirements.txt
flask>=3.1.3,<4.0
flask-login>=0.6.0,<1.0
flask-wtf>=1.2.0,<2.0
werkzeug>=3.1.6,<4.0
argon2-cffi>=21.0.0,<22.0
colormath>=3.0.0,<4.0
pdfplumber>=0.9.0,<1.0
pypdfium2>=5.0.0,<6.0
Pillow>=12.1.1,<13.0
flask-limiter>=3.5.0,<4.0
gunicorn>=22.0.0,<24.0
requests>=2.31.0,<3.0
pyinstaller>=6.0.0,<7.0
